+++ projects/egeo/src/lib/barrels.ts
@@ -45,7 +45,6 @@ import { StTagInputModule } from './st-tag-input/st-tag-input.module';
 import { SdsTextareaModule } from './sds-textarea/sds-textarea.module';
 import { SdsToggleButtonModule } from './sds-toggle-button/sds-toggle-button.module';
 import { StTooltipModule } from './st-tooltip/st-tooltip.module';
-import { StTwoListSelectionModule } from './st-two-list-selection/st-two-list-selection.module';
 import { StVerticalTabsModule } from './st-vertical-tabs/st-vertical-tabs.module';
 import { StFormFieldModule } from './st-form/st-form-field/st-form-field.module';
 import { SdsNotificationModule } from './sds-notification/sds-notification.module';
@@ -93,7 +92,6 @@ export const DECLARATIONS: any[] = [
    SdsTextareaModule,
    SdsToggleButtonModule,
    StTooltipModule,
-   StTwoListSelectionModule,
    StVerticalTabsModule,
    StZeroPageModule
 ];