+++ projects/sds-demo/src/app/demos/barrels.ts
@@ -44,7 +44,6 @@ import { StTagInputDemoModule } from './st-tag-input-demo/st-tag-input-demo.modu
 import { SdsTextareaDemoModule } from './sds-textarea-demo/sds-textarea-demo.module';
 import { SdsToggleButtonDemoModule } from './sds-toggle-button-demo/sds-toggle-button-demo.module';
 import { StTooltipDemoModule } from './st-tooltip-demo/st-tooltip-demo.module';
-import { StTwoListSelectionDemoModule } from './st-two-list-selection-demo/st-two-list-selection-demo.module';
 import { StVerticalTabsDemoModule } from './st-vertical-tabs-demo/st-vertical-tabs-demo.module';
 import { StFilterSelectorDemoModule } from './st-filter-selector-demo/st-filter-selector-demo.module';
 import { StZeroPageDemoModule } from './st-zero-page-demo/st-zero-page-demo.module';
@@ -93,7 +92,6 @@ export const EGEO_DEMO_MODULES: any[] = [
    SdsTextareaDemoModule,
    SdsToggleButtonDemoModule,
    StTooltipDemoModule,
-   StTwoListSelectionDemoModule,
    StVerticalTabsDemoModule,
    StZeroPageDemoModule
 ];